+++ CHANGES.rst
@@ -1,8 +1,8 @@
 Change Log
 ==========
 
-4.0.0bX
--------
+4.0.0rc1
+--------
 
 PRAW 4 introduces significant breaking changes. The numerous changes are not
 listed here, only the feature removals. Please read through